# runner/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Livepeer AI Runner",
        "description": "An application to run AI pipelines",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "https://dream-gateway.livepeer.cloud",
            "description": "Livepeer Cloud Community Gateway"
        }
    ],
    "paths": {
...
        "/llm-generate": {
            "post": {
                "summary": "Llm Generate",
                "operationId": "llm_generate",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_llm_generate_llm_generate_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LlmResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "APIError": {
                "properties": {
                    "msg": {
                        "type": "string",
                        "title": "Msg"
                    }
                },
                "type": "object",
                "required": [
                    "msg"
                ],
                "title": "APIError"
            },
...
            "Body_llm_generate_llm_generate_post": {
                "properties": {
                    "prompt": {
                        "type": "string",
                        "title": "Prompt"
                    },
                    "model_id": {
                        "type": "string",
                        "title": "Model Id",
                        "default": ""
                    },
                    "system_msg": {
                        "type": "string",
                        "title": "System Msg"
                    },
                    "temperature": {
                        "type": "number",
                        "title": "Temperature"
                    },
                    "max_tokens": {
                        "type": "integer",
                        "title": "Max Tokens"
                    }
                },
                "type": "object",
                "required": [
                    "prompt"
                ],
                "title": "Body_llm_generate_llm_generate_post"
            },
...
            "HTTPError": {
                "properties": {
                    "detail": {
                        "$ref": "#/components/schemas/APIError"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "HTTPError"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LlmResponse": {
                "properties": {
                    "response": {
                        "type": "string",
                        "title": "Response"
                    },
                    "tokens_used": {
                        "type": "integer",
                        "title": "Tokens Used"
                    }
                },
                "type": "object",
                "required": [
                    "response",
                    "tokens_used"
                ],
                "title": "LlmResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}